third_party/blink/web_tests/platform/win/external/wpt/css/css-fonts/variations/at-font-face-font-matching-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Matching font-style: 'oblique 20deg' should prefer 'oblique -40deg -30deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 21deg' should prefer 'oblique -40deg -30deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 10deg' should prefer 'oblique 40deg 50deg' over 'oblique 0deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 10deg' should prefer 'oblique -40deg -30deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique 0deg' should prefer 'oblique -40deg -30deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -10deg' should prefer 'oblique -50deg -40deg' over 'oblique 0deg 10deg'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -10deg' should prefer 'oblique 40deg 50deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -20deg' should prefer 'oblique 40deg 50deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
[FAIL] Matching font-style: 'oblique -21deg' should prefer 'oblique 40deg 50deg' over 'italic'
  assert_equals: Unexpected font on test element expected 487 but got 532
Harness: the test ran to completion.
